+++ blueprints/flexberry-edit-form/index.js
@@ -82,6 +82,7 @@ module.exports = {
             parentRoute: editFormBlueprint.parentRoute,
             flexberryComponents: editFormBlueprint.flexberryComponents,
             functionGetCellComponent: editFormBlueprint.functionGetCellComponent,
+            isEmberCpValidationsUsed: editFormBlueprint.isEmberCpValidationsUsed,
         }, editFormBlueprint.locales.getLodashVariablesProperties() // for use in files\__root__\locales\**\forms\__name__.js
         );
     }
@@ -90,6 +91,7 @@ var EditFormBlueprint = /** @class */ (function () {
     function EditFormBlueprint(blueprint, options) {
         this.snippetsResult = [];
         this._tmpSnippetsResult = [];
+        this.isEmberCpValidationsUsed = ModelBlueprint_1.default.checkCpValidations(blueprint);
         this.blueprint = blueprint;
         this.options = options;
         this.modelsDir = path.join(options.metadataDir, "models");
@@ -117,7 +119,8 @@ var EditFormBlueprint = /** @class */ (function () {
         }
     }
     EditFormBlueprint.prototype.readSnippetFile = function (fileName, fileExt) {
-        return stripBom(fs.readFileSync(path.join(this.blueprint.path, "snippets", fileName + "." + fileExt), "utf8"));
+        var snippetsFolder = this.isEmberCpValidationsUsed ? "ember-cp-validations" : "ember-validations";
+        return stripBom(fs.readFileSync(path.join(this.blueprint.path, "snippets", snippetsFolder, fileName + "." + fileExt), "utf8"));
     };
     EditFormBlueprint.prototype.readHbsSnippetFile = function (componentName) {
         return this.readSnippetFile(componentName, "hbs");
@@ -164,6 +167,7 @@ var EditFormBlueprint = /** @class */ (function () {
             projAttr.type = attr.type;
             projAttr.entityName = this.options.entity.name;
             projAttr.dashedName = (projAttr.name || '').replace(/\./g, '-');
+            this.calculateValidatePropertyNames(projAttr);
             this._tmpSnippetsResult.push({ index: projAttr.index, snippetResult: lodash.template(snippet)(projAttr) });
         }
         this.fillBelongsToAttrs(proj.belongsTo, []);
@@ -180,6 +184,8 @@ var EditFormBlueprint = /** @class */ (function () {
                 belongsTo.readonly = "readonly";
                 belongsTo.entityName = this.options.entity.name;
                 belongsTo.dashedName = (belongsTo.name || '').replace(/\./g, '-');
+                belongsTo.isDropdown = belongsTo.type === 'combo';
+                this.calculateValidatePropertyNames(belongsTo);
                 this._tmpSnippetsResult.push({ index: belongsTo.index, snippetResult: lodash.template(this.readHbsSnippetFile("flexberry-lookup"))(belongsTo) });
             }
         }
@@ -191,6 +197,7 @@ var EditFormBlueprint = /** @class */ (function () {
             hasMany.entityName = this.options.entity.name;
             hasMany.dashedName = (hasMany.name || '').replace(/\./g, '-');
             this.locales.setupEditFormAttribute(hasMany);
+            this.calculateValidatePropertyNames(hasMany);
             this.snippetsResult.push(lodash.template(this.readHbsSnippetFile("flexberry-groupedit"))(hasMany));
         }
     };
@@ -213,6 +220,7 @@ var EditFormBlueprint = /** @class */ (function () {
                 belongsToAttr.entityName = this.options.entity.name;
                 belongsToAttr.dashedName = (belongsToAttr.name || '').replace(/\./g, '-');
                 this.locales.setupEditFormAttribute(belongsToAttr);
+                this.calculateValidatePropertyNames(belongsToAttr);
                 this._tmpSnippetsResult.push({ index: belongsToAttr.index, snippetResult: lodash.template(snippet)(belongsToAttr) });
             }
             this.fillBelongsToAttrs(belongsTo.belongsTo, currentPath);
@@ -242,6 +250,13 @@ var EditFormBlueprint = /** @class */ (function () {
             targetRoot = isDummy ? path.join("tests/dummy", targetRoot) : "addon";
         }
         return lodash.template(path.join(targetRoot, "locales", "${ locale }", localePathSuffix));
+    };
+    EditFormBlueprint.prototype.calculateValidatePropertyNames = function (attrs) {
+        var name = attrs.name;
+        var lastDotIndex = name.lastIndexOf(".");
+        var isHaveMaster = lastDotIndex > 0 && lastDotIndex < (name.length - 1);
+        attrs.propertyMaster = (isHaveMaster) ? "." + name.substring(0, lastDotIndex) : "";
+        attrs.propertyName = (isHaveMaster) ? name.substring(lastDotIndex + 1, name.length) : name;
     };
     return EditFormBlueprint;
 }());