+++ tests/dummy/app/controllers/components-acceptance-tests/flexberry-objectlistview/custom-filter.js
@@ -45,7 +45,7 @@ export default ListFormController.extend({
 
     componentForFilter(type, relation) {
       switch (type) {
-        case 'date': return { name: 'flexberry-datepicker' };
+        case 'date': return { name: 'flexberry-simpledatetime', properties: { type: 'date' } };
         case 'decimal': return { name: 'flexberry-textbox', properties: { class: 'compact fluid' } };
         default: return {};
       }